requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
spotipy==2.22.1
urllib3==2.0.7